Got Jump working on Captain Combo
Added an Empty Child GameObject to Captain Combo for Ground Detection. Had to replace the ground check of parent to target child.
Made fixes to the rigidbody 2D component in Combo and modified it as Never Sleep and Continuous and Interpolate.

For input info, Go to Edit --> Project Settings --> Input Manager. The visual script should have the matching input tag to that on the input manager list.