Dev Test Room
Added a Dev Testing Room for:
- All functionality of enemy prefabs
- All functionality of player and companion prefabs
- Game mechanics for later progression of final product.

Updated the Kalhorn script to contain functionality of Enemy Navigation and fully support Animation State Machine.
In editor, go to Dev Room scene, PLAY the editor, go to the Kalhorn prefab, and in the "Kalhorn script", uncheck isAlive and this'll demonstrate that the animator state machine is reacting to the setbool function for when enemy is dead. A thread delay is in effect until it destroys itself from the active scene. For each enemy, functionality is a little different, but easy enough to narrow what each enemy will do.